fix: make createAttachments abstract and fix ms graph implementation
